Revert "feat: add solana chains"
