+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from "./evm";
 export * from "./helpers";
 export * from "./sablier";
 export * from "./types";
