--- src/solana/chains/index.ts
@@ -1,2 +1,0 @@
-export * as chains from "./data";
-export * from "./data";
