--- src/solana/chains/queries.ts
@@ -1,5 +1,0 @@
-import { createChainQueries } from "@src/internal/factories/chains";
-import type { Sablier } from "@src/types";
-import * as chains from "./data";
-
-export const chainsQueries = createChainQueries<Sablier.Solana.Chain>(chains);
