--- src/solana/enums.ts
@@ -1,15 +1,0 @@
-export enum ChainCode {
-  mainnet = "SOL",
-  devnet = "SOLDEV",
-  testnet = "SOLTEST",
-}
-export enum Cluster {
-  mainnet = "mainnet-beta",
-  devnet = "devnet",
-  testnet = "testnet",
-}
-
-export const enums = {
-  ChainCode,
-  Cluster,
-};
