--- src/solana/index.ts
@@ -1,1 +1,0 @@
-export { chains } from "./chains";
